faceai/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
mediapipe==0.10.11
opencv-python
numpy
pydantic
torch
torchvision
pillow
